requirements.txt
# For plotting
dash==1.3.0  # The core dash backend
dash-daq==0.2.1
plotly-geo==1.0.0
dash-cytoscape==0.0.5

# Nice JSON encoder / decoder
python-rapidjson==0.8.0

psycopg2-binary==2.8.3
sqlalchemy==1.3.10
sqlalchemy-utils==0.34.2